Switch to doubled hex coordinate system

+++ bit-array.c
@@ -15,7 +15,7 @@ void bit_array_destroy(BitArray *b)
 i8 bit_array_has(BitArray *b, i32 i)
 {
 	u64 bit = (u64) 1 << (i % 64);
-	return (b[i / 64] & bit) != 0;
+	return (b[i / 64] & bit) != (u64) 0;
 }
 
 void bit_array_set(BitArray *b, i32 i)
@@ -46,6 +46,6 @@ i32 bit_array_count_has(BitArray *b, i32 capacity)
 // Set the bit if src has bit set, but don't clear if it does not.
 void bit_array_copy_set(BitArray *dest, i32 dest_i, BitArray *src, i32 src_i)
 {
-	u64 dest_shift = dest_i % 64;
+	i32 dest_shift = dest_i % 64;
 	dest[dest_i / 64] |= (u64) bit_array_has(src, src_i) << dest_shift;
 }
